+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/BrowserToolbar.kt
@@ -198,6 +198,10 @@ class BrowserToolbar @JvmOverloads constructor(
     }
 
     override fun setSearchTerms(searchTerms: String) {
+        if (state == State.EDIT) {
+            edit.editSuggestion(searchTerms)
+        }
+
         this.searchTerms = searchTerms
     }